# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Warehouse Management API
  description: |
    API for warehouse operations, stock, approvals, reporting, and exports.

    Request and response JSON property names use snake_case. Successful responses use the common
    envelope `{success, code, message, data, errors}`. Errors use the same envelope with `success=false`,
    `data=null`, and field errors keyed by snake_case property names. Bean validation failures return 422.

    Paginated list endpoints accept optional `page` (default 1), `size` (default 10, maximum 100),
    `sort` (endpoint-supported field), `order` (`asc` or `desc`, default `desc`), and `search`.
    Authentication uses a bearer JWT. Inbound, outbound, transfer, and adjustment records follow
    draft -> submit -> approve/reject workflows; approval and rejection require their corresponding
    permission, and creators cannot review their own records. Workflow conflicts return 409 with a
    stable error code such as `*_NOT_REVIEWABLE`, `*_NOT_EDITABLE`, or `INSUFFICIENT_STOCK`; a creator
    attempting self-review receives 403 with a `*_SELF_REVIEW_FORBIDDEN` code.
  version: 1.0.0
servers:
- url: http://localhost:8081
  description: Generated server url
...
paths:
...
  /api/v1/reports/stocks:
    get:
      tags:
      - Reports
      operationId: stocks
      parameters:
      - name: warehouseId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: page
        in: query
        description: One-based page number.
        required: false
        schema:
          type: integer
          format: int32
          default: 1
          description: One-based page number.
          minimum: 1
      - name: size
        in: query
        description: Number of records per page (1-100).
        required: false
        schema:
          type: integer
          format: int32
          default: 10
          description: Number of records per page (1-100).
          maximum: 100
          minimum: 1
      - name: sort
        in: query
        description: Field used to sort the results.
        required: false
        schema:
          type: string
          default: created_at
          description: Field used to sort the results.
      - name: order
        in: query
        description: "Sort direction: asc or desc."
        required: false
        schema:
          type: string
          default: desc
          description: "Sort direction: asc or desc."
          enum:
          - asc
          - desc
          pattern: (?i)^(asc|desc)$
      - name: search
        in: query
        description: Optional text filter.
        required: false
        schema:
          type: string
          description: Optional text filter.
          maxLength: 100
          minLength: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseStockBalanceResponse"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
  /api/v1/reports/stocks/summary:
    get:
      tags:
      - Reports
      operationId: stockSummary
      parameters:
      - name: warehouseId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseStockSummaryResponse"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
  /api/v1/reports/movements:
    get:
      tags:
      - Reports
      operationId: movements
      parameters:
      - name: warehouseId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: fromDate
        in: query
        description: Inclusive start date in UTC.
        required: false
        schema:
          type: string
          format: date
          description: Inclusive start date in UTC.
      - name: toDate
        in: query
        description: Inclusive end date in UTC.
        required: false
        schema:
          type: string
          format: date
          description: Inclusive end date in UTC.
      - name: page
        in: query
        description: One-based page number.
        required: false
        schema:
          type: integer
          format: int32
          default: 1
          description: One-based page number.
          minimum: 1
      - name: size
        in: query
        description: Number of records per page (1-100).
        required: false
        schema:
          type: integer
          format: int32
          default: 10
          description: Number of records per page (1-100).
          maximum: 100
          minimum: 1
      - name: sort
        in: query
        description: Field used to sort the results.
        required: false
        schema:
          type: string
          default: created_at
          description: Field used to sort the results.
      - name: order
        in: query
        description: "Sort direction: asc or desc."
        required: false
        schema:
          type: string
          default: desc
          description: "Sort direction: asc or desc."
          enum:
          - asc
          - desc
          pattern: (?i)^(asc|desc)$
      - name: search
        in: query
        description: Optional text filter.
        required: false
        schema:
          type: string
          description: Optional text filter.
          maxLength: 100
          minLength: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseStockMovementResponse"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
...
  /api/v1/inventory/movements:
    get:
      tags:
      - Inventory
      operationId: movements_1
      parameters:
      - name: page
        in: query
        description: One-based page number.
        required: false
        schema:
          type: integer
          format: int32
          default: 1
          description: One-based page number.
          minimum: 1
      - name: size
        in: query
        description: Number of records per page (1-100).
        required: false
        schema:
          type: integer
          format: int32
          default: 10
          description: Number of records per page (1-100).
          maximum: 100
          minimum: 1
      - name: sort
        in: query
        description: Field used to sort the results.
        required: false
        schema:
          type: string
          default: created_at
          description: Field used to sort the results.
      - name: order
        in: query
        description: "Sort direction: asc or desc."
        required: false
        schema:
          type: string
          default: desc
          description: "Sort direction: asc or desc."
          enum:
          - asc
          - desc
          pattern: (?i)^(asc|desc)$
      - name: search
        in: query
        description: Optional text filter.
        required: false
        schema:
          type: string
          description: Optional text filter.
          maxLength: 100
          minLength: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseStockMovementResponse"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    ApiResponsePageResponseStockBalanceResponse:
      type: object
      properties:
        success:
          type: boolean
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/PageResponseStockBalanceResponse"
        errors:
          type: object
          additionalProperties:
            type: string
    PageResponseStockBalanceResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/StockBalanceResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total_elements:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int32
    StockBalanceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        warehouse_id:
          type: string
          format: uuid
        warehouse_code:
          type: string
        warehouse_name:
          type: string
        warehouse_location_id:
          type: string
          format: uuid
        location_code:
          type: string
        location_name:
          type: string
        product_id:
          type: string
          format: uuid
        sku:
          type: string
        product_name:
          type: string
        unit:
          type: string
        quantity:
          type: integer
          format: int32
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ApiResponseStockSummaryResponse:
      type: object
      properties:
        success:
          type: boolean
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/StockSummaryResponse"
        errors:
          type: object
          additionalProperties:
            type: string
    StockSummaryResponse:
      type: object
      properties:
        warehouse_id:
          type: string
          format: uuid
        total_products:
          type: integer
          format: int64
        total_stock:
          type: integer
          format: int64
        low_stock_products:
          type: integer
          format: int64
        out_of_stock_products:
          type: integer
          format: int64
    ApiResponsePageResponseStockMovementResponse:
      type: object
      properties:
        success:
          type: boolean
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/PageResponseStockMovementResponse"
        errors:
          type: object
          additionalProperties:
            type: string
    PageResponseStockMovementResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/StockMovementResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total_elements:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int32
    StockMovementResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        warehouse_id:
          type: string
          format: uuid
        warehouse_code:
          type: string
        warehouse_name:
          type: string
        warehouse_location_id:
          type: string
          format: uuid
        location_code:
          type: string
        product_id:
          type: string
          format: uuid
        sku:
          type: string
        product_name:
          type: string
        movement_type:
          type: string
        movement_direction:
          type: string
        quantity:
          type: integer
          format: int32
        stock_after:
          type: integer
          format: int32
        source_entity_type:
          type: string
        source_entity_id:
          type: string
          format: uuid
        source_item_id:
          type: string
          format: uuid
        occurred_at:
          type: string
          format: date-time
        created_by_user_id:
          type: string
          format: uuid
...
    ApiErrorEnvelope:
      type: object
      properties:
        success:
          default: false
        code:
          type: integer
          format: int32
        message:
          type: string
        data: {}
        errors:
          type: object
          additionalProperties:
            type: string
      required:
      - code
      - data
      - errors
      - message
      - success
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT